+++ src/CCharter.cpp
@@ -328,7 +328,7 @@ void CCharter::preRender() {
             }
         }
 
-        for (int i = 0; i < 5; i++) {
+        for (size_t i = 0; i < 5; i++) {
             if (!addedNow) {
                 if (!fretKeys[i] && processingP.fretsPressed[i]) {
                     result->type ^= CGamePlay::notesFlagsConst[i];
